Sprint 2 US #38 Task #44, #48 and #52
Task #44 succesfully completed added a new edit button in my_posts.html.
Task #48 currently in progress now the title and description of posts is shown in editable text area, currently working on saving the changes to the database.
Task #52 Currently in progress timestamp will be added alongwith the changes made to posts in edit mode.
Also added the edit button in the edit_post.html to submit the changes to database.